refactored player interaction with objects and added some soda logic
