+++ movieTheaterSim/Assets/Scripts/Computer/DesktopControllerHandler.cs
@@ -3,14 +3,7 @@ using UnityEngine;
 
 public class DesktopControllerHandler : MonoBehaviour
 {
-    [SerializeField] private Transform PCViewPos; //position of point where player gets moved to   
-    [SerializeField] private Texture2D hoverCursor; //mouse cursor icon that gets switch when viewing pc
     [SerializeField] private TextMeshProUGUI clockText; //clock text in taskbar - TODO: set to gameTime player
-    [SerializeField] private float anglePerSecond = 25.0f; //used to control the speed at which the camera turns at pc
-    [SerializeField] private float camMoveSpeed = 3f; //rate at which the cam moves to the object
-
-    private GameObject player; //the interacting player object
-    private bool isCameraMoving = false;
 
     [Header("Button Gameobject to fetch by name")]
     [SerializeField] private GameObject pcIcons;
@@ -20,65 +13,16 @@ public class DesktopControllerHandler : MonoBehaviour
     [SerializeField] private GameObject bankApp;
     [SerializeField] private GameObject testApp3;
 
-    private bool canInteract = false;
+    private void Start()
+    {
+        clockText = this.transform.Find("TaskBarBackground/TaskBarTime_(TMP)").GetComponent<TextMeshProUGUI>();
+    }
 
     // Update is called once per frame
-    void Update()
+    private void Update()
     {
-        if (Input.GetButtonDown("Interaction") && canInteract && player != null)
-        {
-            if (player.GetComponent<FirstPersonController>().enabled == false)
-            {
-                //stops the update of moving to target
-                isCameraMoving = false;
-                //resets the local pos of camera back to player, since orginal is 0.0.0 under first person controller
-                player.GetComponent<FirstPersonController>().playerCamera.transform.localPosition = new Vector3(0, 0, 0);
-
-                //allows the first person controller to move again
-                Cursor.lockState = CursorLockMode.Locked;
-                player.GetComponent<FirstPersonController>().enabled = true;
-            }
-            else
-            {
-                //turns off movement for first person controller character
-                Cursor.lockState = CursorLockMode.Confined;
-                player.GetComponent<FirstPersonController>().enabled = false;
-                //turns on movement of player camera to target pos              
-                isCameraMoving = true;
-            }
-        }
-
-        if (isCameraMoving && player != null)
-        {
-            //moves to target location smoothly over time
-            player.GetComponent<FirstPersonController>().playerCamera.transform.position = Vector3.MoveTowards(player.GetComponent<FirstPersonController>().playerCamera.transform.position, PCViewPos.transform.position, camMoveSpeed * Time.deltaTime);
-
-            //Rotates to target location smoothly over time
-            player.GetComponent<FirstPersonController>().playerCamera.transform.rotation = Quaternion.RotateTowards(player.GetComponent<FirstPersonController>().playerCamera.transform.rotation, PCViewPos.transform.rotation, anglePerSecond * Time.deltaTime);
-
-        }
-
         //change this later to actual game time
         clockText.text = System.DateTime.Now.ToString("h:mm tt");
-    }
-
-    private void OnTriggerEnter(Collider other)
-    {
-       if(other.gameObject.CompareTag("Player"))
-       {
-            canInteract = true;
-            player = other.gameObject;
-        }
-  
-    }
-
-    private void OnTriggerExit(Collider other)
-    {
-        if (other.gameObject.CompareTag("Player"))
-        {
-            canInteract = false;
-            player = null;
-        }
     }
 
     /// |*************************| UI LOGIC |*************************|
@@ -93,11 +37,13 @@ public class DesktopControllerHandler : MonoBehaviour
         ClickOnIcon();
         wholeSaleApp.SetActive(true);
     }
+
     public void BankAppIconOnClick()
     {
         ClickOnIcon();
         bankApp.SetActive(true);
     }
+
     public void TestApp3OnClick()
     {
         ClickOnIcon();
