+++ movieTheaterSim/Assets/Scripts/PlayerInteractions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2d6f1a3962ec000478fe40351f2ced2d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
