+++ movieTheaterSim/Assets/Scripts/PlayerInteractions/PlayerObjectInteractionHandler.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ffa99000cb3237d42b9f784dad490dca
