+++ movieTheaterSim/Assets/Scripts/SodaMachine/SodaSlotHandler.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+public class SodaSlotHandler : MonoBehaviour
+{
+    [SerializeField] public string sodaName;
+    [SerializeField] public int sodaFuleLevel;
+
+}
