+++ movieTheaterSim/Assets/Scripts/SodaMachine/SodaSlotHandler.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 69780a3939adbfd4f86aacc2585e272c
